chore: move botcontroller into game
